requirements.txt
numpy==1.23.5
mmcv_full==1.7.1
mmdet==2.28.2
mmsegmentation==0.30.0
urllib3==1.26.16
pyquaternion==0.9.9
nuscenes-devkit==1.1.10
yapf==0.33.0
tensorboard==2.14.0
motmetrics==1.1.3
pandas==1.1.5
